Bluetooth MAP (Message Access Profile) Upstream Changes (1/3)
PORTED FROM CM10: http://review.cyanogenmod.org/#/c/26392/

Incorporates updates to the Bluetooth MAP profile from Code Aurora's
Android Enablement Project since we initially pulled in from their
gingerbread to CM7 last year. This patch addresses bugs with the
current MAP profile, adds a UI request for message access approval
when pairing, and adds additional error checking and logging.
Confirmed to fix date and "Download" function of Ford SYNC Gen1.
Pulled in from the codeaurora jb branch.


Conflicts:
	AndroidManifest.xml

+++ src/com/android/bluetooth/map/MapUtils/MsgListingConsts.java
@@ -30,7 +30,7 @@ package com.android.bluetooth.map.MapUtils;
 
 public class MsgListingConsts {
 
-	public class MsgInfo {
+    public class MsgInfo {
         public String dateTime = null;
 
         public String getDateTime() {
@@ -44,7 +44,7 @@ public class MsgListingConsts {
     }
     public MsgInfo msgInfo = new MsgInfo();
 
-    public int msg_handle = 0;
+    public long msg_handle = 0;
     public String subject = null;
     public boolean  sendSubject = false;
     public String datetime = null;
@@ -64,11 +64,11 @@ public class MsgListingConsts {
     public String msg_protected = null;
     public String replyto_addressing = null;
 
-    public int getMsg_handle() {
+    public long getMsg_handle() {
         return msg_handle;
     }
 
-    public void setMsg_handle(int msg_handle) {
+    public void setMsg_handle(long msg_handle) {
         this.msg_handle = msg_handle;
     }
